Move Sha256Hash and other Hash algorithm types from KeyVault to Azure::Core.
