+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/signature_algorithm.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <azure/keyvault/common/internal/sha_hash.hpp>
+#include <azure/core/internal/cryptography/sha_hash.hpp>
 
 #include "../private/key_constants.hpp"
 #include "azure/keyvault/keys/cryptography/signature_algorithm.hpp"
@@ -28,19 +28,19 @@ namespace Azure {
     if (*this == SignatureAlgorithm::RS256 || *this == SignatureAlgorithm::PS256
         || *this == SignatureAlgorithm::ES256 || *this == SignatureAlgorithm::ES256K)
     {
-      return std::make_unique<_internal::Sha256Hash>();
+      return std::make_unique<Azure::Core::Cryptography::_internal::Sha256Hash>();
     }
 
     if (*this == SignatureAlgorithm::RS384 || *this == SignatureAlgorithm::PS384
         || *this == SignatureAlgorithm::ES384)
     {
-      return std::make_unique<_internal::Sha384Hash>();
+      return std::make_unique<Azure::Core::Cryptography::_internal::Sha384Hash>();
     }
 
     if (*this == SignatureAlgorithm::RS512 || *this == SignatureAlgorithm::PS512
         || *this == SignatureAlgorithm::ES512)
     {
-      return std::make_unique<_internal::Sha512Hash>();
+      return std::make_unique<Azure::Core::Cryptography::_internal::Sha512Hash>();
     }
     throw std::runtime_error("Unkown Hash algorithm for: " + m_value);
   }
